Complete destroy assignment

+++ app/controllers/comments_controller.rb
@@ -1,43 +1,39 @@
 class CommentsController < ApplicationController
-  
-  before_filter :authenticate_user!
-  
+
   def create
     @post = Post.find(params[:post_id])
-    @comment = current_user.comments.new(comment_params)
+    @comment = current_user.comments.build(comment_params)
     @comment.post = @post
-    @new_comment = Comment.new
+    @topic = @post.topic
 
     authorize @comment
 
     if @comment.save
       flash[:notice] = "Comment was saved."
-      redirect_to @post 
+      redirect_to [@topic, @post]
     else
-      flash[:error] = "Error creating comment. Please try again."
-      render "posts/show"
+      flash[:error] = "There was an error saving the comment. Please try again."
+      redirect_to [@topic, @post]
     end
   end
 
   def destroy
-    @topic = Topic.find(params[:topic_id])
-    @post = @topic.posts.find(params[:post_id])
+    @post = Post.find(params[:post_id])
     @comment = @post.comments.find(params[:id])
- 
     authorize @comment
-    
-     if @comment.destroy
-      flash[:notice] = "Comment was removed."
-      redirect_to [@topic, @post]
-     else
-      flash[:error] = "Comment couldn't be deleted. Try again."
-      redirect_to [@topic, @post]
+
+    authorize @comment
+    if @comment.destroy
+      flash[:notice] = "Comment was deleted successfully."
+    else
+      flash[:error] = "There was an error deleting the comment. Please try again."
     end
+    redirect_to [@post.topic, @post]
   end
 
   private
 
   def comment_params
-    params.require(:comment).permit(:body)
+    params.require(:comment).permit(:body, :post_id)
   end
 end
